Move PC class to a separate module for less cognitive overload
